Move UserModelDataMapper to presentation layer
